+++ R/prepare_solution.R
@@ -5,11 +5,12 @@
 #' @export
 #' @param data A data frame with the following columns:
 #'
-#' * `t` Time \eqn{t} (days).
+#' * `t` Time \eqn{t}{t} (days).
 #' * `ee` Eccentricity \eqn{e} (unitless).
 #' * `lph` Longitude of perihelion \eqn{\varpi} (degrees).
 #' * `lan` Longitude of the ascending node \eqn{\Omega} (degrees).
 #' * `inc` Inclination \eqn{I} (degrees).
+#'
 #' The easiest way to get this is with [get_solution()].
 # inherit quiet
 #' @inheritParams get_ZB
@@ -70,8 +71,8 @@ prepare_solution <- function(data, quiet = FALSE) {
       lphu = unwrap(.data$lph),
       lanu = unwrap(.data$lan)
     ) |>
-    dplyr::mutate(t_ka = .data$t / KY2D, .after = "t") |>
-    dplyr::mutate(age = -.data$t_ka, .after = "t_ka") |>
+    dplyr::mutate(time = .data$t / KY2D, .after = "t") |>
+    ## dplyr::mutate(age = -.data$time, .after = "time") |>
     dplyr::mutate(
       hh = .data$ee * sin(.data$lph / R2D),
       kk = .data$ee * cos(.data$lph / R2D),